removed certain files, and changed some routes
